backend/requirements.txt
Django==3.2.6
djangorestframework==3.12.2
django-cors-headers==3.7.0
